+++ src/modules/slimer-sdk/webpage.js
@@ -723,6 +723,7 @@ function _create(aParentWindow) {
             let domWindowUtils = win.QueryInterface(Ci.nsIInterfaceRequestor)
                                     .getInterface(Ci.nsIDOMWindowUtils);
             domWindowUtils. setCSSViewport(w,h);
+            win.resizeTo(w,h);
         },
 
         get windowName () {